+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
  * added `Symfony\Component\HttpKernel\Controller\ArgumentResolver`
  * added `Symfony\Component\HttpKernel\DataCollector\RequestDataCollector::getMethod()`
  * added `Symfony\Component\HttpKernel\DataCollector\RequestDataCollector::getRedirect()`
+ * added the `kernel.controller_arguments` event, triggered after controller arguments have been resolved
 
 3.0.0
 -----
@@ -24,8 +25,8 @@ CHANGELOG
  * removed `Symfony\Component\HttpKernel\EventListener\RouterListener::setRequest()`
  * removed `Symfony\Component\HttpKernel\EventListener\ProfilerListener::onKernelRequest()`
  * removed `Symfony\Component\HttpKernel\Fragment\FragmentHandler::setRequest()`
- * removed `Symfony\Component\HttpKernel\HttpCache\Esi::hasSurrogateEsiCapability()` 
- * removed `Symfony\Component\HttpKernel\HttpCache\Esi::addSurrogateEsiCapability()` 
+ * removed `Symfony\Component\HttpKernel\HttpCache\Esi::hasSurrogateEsiCapability()`
+ * removed `Symfony\Component\HttpKernel\HttpCache\Esi::addSurrogateEsiCapability()`
  * removed `Symfony\Component\HttpKernel\HttpCache\Esi::needsEsiParsing()`
  * removed `Symfony\Component\HttpKernel\HttpCache\HttpCache::getEsi()`
  * removed `Symfony\Component\HttpKernel\DependencyInjection\ContainerAwareHttpKernel`